feat: Add new color variable for third color in tailwind.config.ts
This commit adds a new color variable called "third" to the tailwind.config.ts file. The "third" color is set to "#ff2626". This variable can be used to style elements with the third color in the application.

feat: Add "cid" property to tokenInfo in page.tsx

This commit adds a new property called "cid" to the tokenInfo object in the page.tsx file. The "cid" property is assigned the value of tokenInfo.cid. This change allows the tokenInfo object to store the "cid" value for further use in the application.

feat: Add "cid" and "cid" properties to BuySellLayout.tsx

This commit adds two new properties, "cid" and "cid", to the BuySellLayout component in the BuySellLayout.tsx file. The "cid" property is assigned the value of cid, and the "cid" property is assigned the value of cid. These changes enable the BuySellLayout component to access and use the "cid" and "cid" values.

feat: Add new color variable for third color in globals.css

This commit adds a new color variable called "third" to the globals.css file. The "third" color is set to "#ff2626". This variable can be used to style elements with the third color in the application.

feat: Update TradeArr component in Tokens.tsx

This commit updates the TradeArr component in the Tokens.tsx file. The component's className is modified to conditionally apply different styles based on a boolean value. Additionally, the component's className is updated to use the "fill-[#86BF77]" class when the boolean value is true, and the "rotate-180 fill-red-500" class when the boolean value is false. These changes allow the TradeArr component to dynamically apply different styles based on the boolean value.

feat: Add new API endpoints for users' trades, total assets, and memes

This commit adds new API endpoints for retrieving users' trades, total assets, and memes. The endpoints are implemented in the apis.ts file and are named UsersTrades, UsersTotalAssets, and UsersMemes, respectively. These new endpoints allow the application to fetch and display users' trade data, total assets, and memes.

feat: Update MypageModal component in MypageModal.tsx

This commit updates the MypageModal component in the MypageModal.tsx file. The component's className is modified to use the "border-secondary" class for the border color. Additionally, the component's className is updated to use the "text-third" class for the text color. These changes improve the visual appearance of the MypageModal component.

feat: Update Header component in Header.tsx

This commit updates the Header component in the Header.tsx file. The component's className is modified to use the "border-secondary" class for the border color. Additionally, the component's className is updated to use the "text-third" class for the text color. These changes enhance the visual style of the Header component.