feat: Increase timeout for setting gifOn to false in Raffle component
Update the timeout duration from 1 second to 2 seconds in the useEffect hook of the Raffle component. This change allows for a longer delay before setting the gifOn state to false.

Refactor the flex height and add responsive styling in BuySellButtonSection component

In the BuySellButtonSection component of the Tradesection module, refactor the flex height and add responsive styling. The flex height is changed from 50px to 10px, and the responsive styling is added for medium-sized screens.

Update text color in TokenDesc component

Change the text color in the TokenDesc component from white to #AEAEAE. This improves the readability and visual consistency of the component.

Add new interfaces for MyTradeResponse in type.ts

Add two new interfaces, MyTradeResponse and TokenResponse, to the type.ts file. These interfaces define the structure of the response objects returned by the API.

Update text color and content in Footer component

Change the text color in the Footer component from #F2D1E4 to #AEAEAE. Also, update the content to reflect the correct company name and rights information.

Add window resize functionality in Header component

Implement window resize functionality in the Header component. The width of the window is stored in the 'width' state variable, and the value is updated whenever the window is resized. This allows for responsive behavior in the component.

Update Inputform component to add validation for minimum value

In the Inputform component, add validation for the minimum value allowed. If the input value is less than 0.01 ETH, the border color and text color will change to indicate an invalid value.

Update SlotLayout component to include winner redirection logic

In the SlotLayout component, add logic to redirect to the appropriate page based on the winner's raffles. If the winner's raffles match the current page's token, the user is redirected to the '/raffle' page. Otherwise, the user is redirected to the '/raffle/fail' page.

Update Create page to handle threshold input value

In the Create page, add handling for the threshold input value. If the input value is less than 0.01, an alert is displayed to indicate an invalid value.

Update TokensLayout component to remove page navigation icons

Remove the page navigation icons (PageFirst and PageLast) from the TokensLayout component. This simplifies the UI and improves the user experience.